Patch debug feature
Previous changes introduced an annoying bug with the debug feature.
The following patch it, organize logger logic in its own file and fix a
few debug messages.
