+++ cmd/root.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
 	"strings"
@@ -10,15 +9,15 @@ import (
 	"github.com/fabio42/ssl-checker/ui"
 
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/rs/zerolog"
+	// "github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
-const (
-	logFile = "./ssl-checker.log"
-)
+// const (
+// 	logFile = "./ssl-checker.log"
+// )
 
 var (
 	configFile string
@@ -31,6 +30,16 @@ var rootCmd = &cobra.Command{
 	Long:  "ssl-checker is a tool to _quickly_ check certificate details of multiple https targets.",
 
 	PersistentPreRun: func(cmd *cobra.Command, args []string) {
+		err := setLogger(viper.GetBool("debug"))
+		if err != nil {
+			log.Fatal().Msgf("Error failed to configure logger:", err)
+		}
+
+		if viper.GetBool("debug") {
+			// zerolog.SetGlobalLevel(zerolog.DebugLevel)
+			log.Warn().Msgf("Debug is enabled, log will be found in %v", logFile)
+		}
+
 		cfgFile := filepath.Base(configFile)
 		cfgPath := filepath.Dir(configFile)
 		viper.SetConfigName(cfgFile[:len(cfgFile)-len(filepath.Ext(cfgFile))])
@@ -44,11 +53,6 @@ var rootCmd = &cobra.Command{
 		}
 	},
 	Run: func(cmd *cobra.Command, args []string) {
-		if viper.GetBool("debug") {
-			zerolog.SetGlobalLevel(zerolog.DebugLevel)
-			log.Warn().Msgf("Debug is enabled, log will be found in ./zeroDebug.log")
-		}
-
 		envQuery := strings.Split(envCheck, ",")
 
 		if viper.IsSet("queries") {
@@ -134,11 +138,6 @@ func sliceContains(s []string, str string) bool {
 }
 
 func init() {
-	err := setLogger()
-	if err != nil {
-		log.Fatal().Msgf("Error failed to configure logger:", err)
-	}
-
 	rootCmd.PersistentFlags().StringVarP(&configFile, "config", "c", "$HOME/.config/ssl-checker/config.yaml", "Configuration file location")
 	rootCmd.PersistentFlags().BoolP("silent", "s", false, "disable ui")
 	rootCmd.PersistentFlags().BoolP("debug", "d", false, "Enable debug log, out will be saved in "+logFile)
@@ -149,6 +148,11 @@ func init() {
 	viper.BindPFlag("debug", rootCmd.PersistentFlags().Lookup("debug"))
 	viper.BindPFlag("timeout", rootCmd.PersistentFlags().Lookup("timeout"))
 
+	// err := setLogger(viper.GetBool("debug"))
+	// if err != nil {
+	// 	log.Fatal().Msgf("Error failed to configure logger:", err)
+	// }
+
 	rootCmd.AddCommand(listEnvs)
 }
 
@@ -156,51 +160,4 @@ func Execute() {
 	if err := rootCmd.Execute(); err != nil {
 		log.Fatal().Msgf("Whoops. There was an error while executing your CLI '%s'", err)
 	}
-}
-
-// https://github.com/rs/zerolog/issues/150
-type LevelWriter struct {
-	io.Writer
-	Level zerolog.Level
-}
-
-func (lw *LevelWriter) WriteLevel(l zerolog.Level, p []byte) (n int, err error) {
-	if l >= lw.Level {
-		return lw.Write(p)
-	}
-	return len(p), nil
-}
-
-func setLogger() error {
-	var logWriter *os.File
-	var err error
-	zerolog.TimeFieldFormat = zerolog.TimeFormatUnix
-	zerolog.SetGlobalLevel(zerolog.InfoLevel)
-	if viper.GetBool("debug") {
-		zerolog.SetGlobalLevel(zerolog.DebugLevel)
-
-		logWriter, err = os.OpenFile(
-			logFile,
-			os.O_APPEND|os.O_CREATE|os.O_WRONLY,
-			0664,
-		)
-		if err != nil {
-			panic(err)
-		}
-	}
-
-	fileWriter := zerolog.New(zerolog.ConsoleWriter{
-		Out:          logWriter,
-		NoColor:      true,
-		PartsExclude: []string{"time", "level"},
-	})
-	consoleWriter := zerolog.NewConsoleWriter(
-		func(w *zerolog.ConsoleWriter) {
-			w.Out = os.Stderr
-			w.PartsExclude = []string{"time"}
-		},
-	)
-	consoleWriterLeveled := &LevelWriter{Writer: consoleWriter, Level: zerolog.InfoLevel}
-	log.Logger = zerolog.New(zerolog.MultiLevelWriter(fileWriter, consoleWriterLeveled)).With().Timestamp().Logger()
-	return nil
 }
